refactor: removing context from file names
