--- src/containers/PodcasterDashboardOnboarding/interfaces.ts
@@ -1,3 +1,0 @@
-export interface PodcasterDashboardOnboardingProps {
-  onOpenDialog: () => void;
-}
