+++ src/containers/PodcasterOnboarding/interfaces.ts
@@ -1,0 +1,3 @@
+export interface PodcasterOnboardingProps {
+  onOpenDialog: () => void;
+}
