+++ src/containers/index.ts
@@ -3,8 +3,8 @@ export { default as AudioPlayer } from "./AudioPlayer";
 export { default as AuthListener } from "./AuthListener";
 export { default as StyledDrawer } from "./StyledDrawer";
 export { default as CategoriesForm } from "./CategoriesForm";
-export { default as PodcasterDashboardAppBar } from "./DashboardAppBar";
-export { default as PodcasterDashboardOverview } from "./PodcasterDashboardOverview";
-export { default as PodcasterCreateSeriesDialog } from "./CreatePodcastDialog";
-export { default as PodcasterDashboardOnboarding } from "./PodcasterDashboardOnboarding";
-export { default as PodcasterEpisodeCreationDialog } from "./PodcasterEpisodeCreationDialog";
+export { default as DashboardAppBar } from "./DashboardAppBar";
+export { default as DashboardOverview } from "./DashboardOverview";
+export { default as CreatePodcastDialog } from "./CreatePodcastDialog";
+export { default as PodcasterOnboarding } from "./PodcasterOnboarding";
+export { default as CreateEpisodeDialog } from "./CreateEpisodeDialog";
